+++ packages/core/src/generators/mitosis/generator.ts
@@ -1,3 +1,4 @@
+import { ToMitosisOptions } from '@/generators/mitosis/types';
 import {
   runPostCodePlugins,
   runPostJsonPlugins,
@@ -6,25 +7,21 @@ import {
 } from '@/modules/plugins';
 import json5 from 'json5';
 import { format } from 'prettier/standalone';
-import { HOOKS } from '../constants/hooks';
-import { SELF_CLOSING_HTML_TAGS } from '../constants/html_tags';
-import { dedent } from '../helpers/dedent';
-import { fastClone } from '../helpers/fast-clone';
-import { getComponents } from '../helpers/get-components';
-import { getRefs } from '../helpers/get-refs';
-import { getStateObjectStringFromComponent } from '../helpers/get-state-object-string';
-import { isRootTextNode } from '../helpers/is-root-text-node';
-import { mapRefs } from '../helpers/map-refs';
-import { renderPreComponent } from '../helpers/render-imports';
-import { checkHasState } from '../helpers/state';
-import { MitosisComponent } from '../types/mitosis-component';
-import { MitosisNode, checkIsForNode } from '../types/mitosis-node';
-import { BaseTranspilerOptions, TranspilerGenerator } from '../types/transpiler';
-import { blockToReact, componentToReact } from './react';
-
-export interface ToMitosisOptions extends BaseTranspilerOptions {
-  format: 'react' | 'legacy';
-}
+import { HOOKS } from '../../constants/hooks';
+import { SELF_CLOSING_HTML_TAGS } from '../../constants/html_tags';
+import { dedent } from '../../helpers/dedent';
+import { fastClone } from '../../helpers/fast-clone';
+import { getComponents } from '../../helpers/get-components';
+import { getRefs } from '../../helpers/get-refs';
+import { getStateObjectStringFromComponent } from '../../helpers/get-state-object-string';
+import { isRootTextNode } from '../../helpers/is-root-text-node';
+import { mapRefs } from '../../helpers/map-refs';
+import { renderPreComponent } from '../../helpers/render-imports';
+import { checkHasState } from '../../helpers/state';
+import { MitosisComponent } from '../../types/mitosis-component';
+import { MitosisNode, checkIsForNode } from '../../types/mitosis-node';
+import { TranspilerGenerator } from '../../types/transpiler';
+import { blockToReact, componentToReact } from '../react';
 
 export const DEFAULT_FORMAT: ToMitosisOptions['format'] = 'legacy';
 
@@ -217,6 +214,8 @@ export const componentToMitosis: TranspilerGenerator<Partial<ToMitosisOptions>> 
       ${json.hooks.onMount.map((hook) => `onMount(() => { ${hook.code} })`)}
 
       ${!json.hooks.onUnMount?.code ? '' : `onUnMount(() => { ${json.hooks.onUnMount.code} })`}
+
+      ${json.style ? `useStyle(\`${json.style}\`)` : ''}
 
       return (${addWrapper ? '<>' : ''}
         ${json.children.map((item) => blockToMitosis(item, options, component)).join('\n')}