+++ packages/core/src/generators/svelte/types.ts
@@ -3,3 +3,5 @@ import { BaseTranspilerOptions } from '@/types/transpiler';
 export type ToSvelteOptions = BaseTranspilerOptions & {
   stateType?: 'proxies' | 'variables';
 };
+
+export type SvelteMetadata = {};