+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @builder.io/mitosis-cli
 
+## 0.4.5
+
+### Patch Changes
+
+- Updated dependencies [428b3ab]
+  - @builder.io/mitosis@0.4.5
+
+## 0.4.4
+
+### Patch Changes
+
+- Updated dependencies [ad7e576]
+- Updated dependencies [52dc749]
+  - @builder.io/mitosis@0.4.4
+
 ## 0.4.3
 
 ### Patch Changes